+++ Changelog.md
@@ -25,6 +25,7 @@ _Changes:_
 - [X] PDO-4568 Kibana: Modify Dashboards to Show Second Region Logs and Metrics
 - [X] PDO-4569 Remove ES, Kibana and Grafana from second region
 - [X] PDO-4610 Retain set value for slack channel alerts
+- [X] PDO-4614 Automate pinning the branch for ping-cloud-dashboards in PCB
 
 ### 1.16.0.0